Update offsets for 1.9.0
Small revisions, notably 50->100 patterns saved.

+++ NHSE.Core/Save/Files/Personal.cs
@@ -164,7 +164,7 @@ namespace NHSE.Core
         public bool ProfileIsMakeVillage
         {
             get => Data[Offsets.ProfileIsMakeVillage] != 0;
-            set => Data[Offsets.ProfileIsMakeVillage] = value ? 1 : 0;
+            set => Data[Offsets.ProfileIsMakeVillage] = (byte)(value ? 1 : 0);
         }
 
         #endregion
